fixing Rule inheritance
