+++ modulecheck-core/src/main/kotlin/modulecheck/core/rule/DisableAndroidResourcesRule.kt
@@ -15,6 +15,8 @@
 
 package modulecheck.core.rule
 
+import modulecheck.api.ModuleCheckRule
+import modulecheck.api.settings.ChecksSettings
 import modulecheck.api.settings.ModuleCheckSettings
 import modulecheck.core.rule.android.UnusedResourcesGenerationFinding
 import modulecheck.parsing.AndroidMcProject
@@ -43,5 +45,9 @@ class DisableAndroidResourcesRule(
     } else {
       emptyList()
     }
+  }
+
+  override fun shouldApply(checksSettings: ChecksSettings): Boolean {
+    return checksSettings.disableAndroidResources
   }
 }
